'modified_v2'

+++ iotProject/src/DBConnect/ImageUpload.java
@@ -8,7 +8,7 @@ import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 
-public class ImageUpload {
+public class ImageUpload {	//	이미지 관련 업로드
 	private static ImageUpload instance = new ImageUpload();
 	
 	public static ImageUpload getInstance() {
@@ -16,13 +16,13 @@ public class ImageUpload {
 	}
 	
 	private DBConnector dbc = new DBConnector();	//	DBConnector 객체생성
-	private Connection conn;	//	db에 접근하게 해주는 객체
-	private String sql = ""; 	//	쿼리1(MariaDB에 들어갈 명령어지문)
+	private Connection conn;
+	private String sql = "";
 	private String sql2 = "";
-	private PreparedStatement pstmt;		//	db에 sql문을 전달해주는 객체
+	private PreparedStatement pstmt;
 	private PreparedStatement pstmt2;
-	private ResultSet rs;	//	db에서 쿼리의 실행결과를 가져오는 객체
-	private String returns; //메소드 성공 여부 반환
+	private ResultSet rs;
+	private String returns;
 	private int result;
 	
 	//이미지 파일 참고 - https://aristatait.tistory.com/16?category=672398, https://hks003.tistory.com/11
@@ -36,21 +36,23 @@ public class ImageUpload {
 			pstmt = conn.prepareStatement(sql);
 			pstmt.setInt(1, number);
 			rs = pstmt.executeQuery();
+			
 			if(rs.next()) {	//	이미지가 있을 때
 				sql2 = "update organization set organization_image=? where number=?"; // organization 테이블의 이미지 업데이트
-				pstmt2 = conn.prepareStatement(sql);
+				pstmt2 = conn.prepareStatement(sql2);
 				pstmt2.setBinaryStream(1, fin, (int)imgFile.length());	//	데이터 베이스에 이미지 파일 업로드
 				pstmt2.setInt(2, number);
 				pstmt2.executeUpdate(); // db에 쿼리문 입력						
+				returns = "orgModified";
 			}
 			else {	//	이미지가 없을 때
 				sql2 = "insert into organization (number, organizaion_image) values (?, ?)"; // organization 테이블의 이미지 넣기
-				pstmt2 = conn.prepareStatement(sql);
+				pstmt2 = conn.prepareStatement(sql2);
 				pstmt2.setInt(1, number);
 				pstmt2.setBinaryStream(2, fin, (int)imgFile.length());	//	데이터 베이스에 이미지 파일 업로드
 				pstmt2.executeUpdate(); // db에 쿼리문 입력				
+				returns = "orgUploaded";
 			}			
-			returns = "orgUploaded";
 		} catch (Exception e) {
 			e.printStackTrace();
 			returns = "error";
@@ -67,11 +69,20 @@ public class ImageUpload {
 		try {
 			Class.forName("org.mariadb.jdbc.Driver");
 			conn = DriverManager.getConnection(dbc.getURL(), dbc.getID(), dbc.getPW()); // 데이터베이스 접근을 위한 로그인
-			sql = "delete from organization where number=?"; // organization테이블의 number에 해당되는 레코드 삭제
+			sql = "select * from organization where number=?";
 			pstmt = conn.prepareStatement(sql);
 			pstmt.setInt(1, number);
-			pstmt.executeQuery(); // db에 쿼리문 입력
-			returns = "orgDeleted";
+			rs = pstmt.executeQuery();
+			if(rs.next()) {
+				sql2 = "delete from organization where number=?"; // organization테이블의 number에 해당되는 레코드 삭제
+				pstmt2 = conn.prepareStatement(sql2);
+				pstmt2.setInt(1, number);
+				pstmt2.executeQuery(); // db에 쿼리문 입력
+				returns = "orgDeleted";				
+			}
+			else {
+				returns = "orgNotExist";
+			}
 		} catch (Exception e) {
 			e.printStackTrace();
 			returns = "error";
@@ -96,19 +107,20 @@ public class ImageUpload {
 			rs = pstmt.executeQuery();
 			if(rs.next()) {	//	이미지가 있을 때
 				sql2 = "update structure set structure_image=? where number=?"; // structure 테이블의 이미지 업데이트
-				pstmt2 = conn.prepareStatement(sql);
+				pstmt2 = conn.prepareStatement(sql2);
 				pstmt2.setBinaryStream(1, fin, (int)imgFile.length());	//	데이터 베이스에 이미지 파일 업로드
 				pstmt2.setInt(2, number);
-				pstmt2.executeUpdate(); // db에 쿼리문 입력						
+				pstmt2.executeUpdate(); // db에 쿼리문 입력
+				returns = "strModified";
 			}
 			else {	//	이미지가 없을 때
 				sql2 = "insert into structure (number, structure_image) values (?, ?)"; // structure 테이블의 이미지 넣기
-				pstmt2 = conn.prepareStatement(sql);
+				pstmt2 = conn.prepareStatement(sql2);
 				pstmt2.setInt(1, number);
 				pstmt2.setBinaryStream(2, fin, (int)imgFile.length());	//	데이터 베이스에 이미지 파일 업로드
 				pstmt2.executeUpdate(); // db에 쿼리문 입력				
+				returns = "strUploaded";
 			}	
-			returns = "strUploaded";
 		} catch (Exception e) {
 			e.printStackTrace();
 			returns = "error";
@@ -125,11 +137,20 @@ public class ImageUpload {
 		try {
 			Class.forName("org.mariadb.jdbc.Driver");
 			conn = DriverManager.getConnection(dbc.getURL(), dbc.getID(), dbc.getPW()); // 데이터베이스 접근을 위한 로그인
-			sql = "delete from structure where number=?"; // structure 테이블의 number에 해당되는 레코드 삭제
+			sql = "select * from structure where number=?";
 			pstmt = conn.prepareStatement(sql);
 			pstmt.setInt(1, number);
-			pstmt.executeQuery(); // db에 쿼리문 입력
-			returns = "strDeleted";
+			rs = pstmt.executeQuery();
+			if(rs.next()) {
+				sql2 = "delete from structure where number=?"; // organization테이블의 number에 해당되는 레코드 삭제
+				pstmt2 = conn.prepareStatement(sql2);
+				pstmt2.setInt(1, number);
+				pstmt2.executeQuery(); // db에 쿼리문 입력
+				returns = "strDeleted";				
+			}
+			else {
+				returns = "strNotExist";
+			}
 		} catch (Exception e) {
 			e.printStackTrace();
 			returns = "error";
@@ -154,19 +175,21 @@ public class ImageUpload {
 			rs = pstmt.executeQuery();
 			if(rs.next()) {	//	이미지가 있을 때
 				sql2 = "update ip set ip_image=? where number=?"; // ip 테이블의 이미지 업데이트
-				pstmt2 = conn.prepareStatement(sql);
+				pstmt2 = conn.prepareStatement(sql2);
 				pstmt2.setBinaryStream(1, fin, (int)imgFile.length());	//	데이터 베이스에 이미지 파일 업로드
 				pstmt2.setInt(2, number);
-				pstmt2.executeUpdate(); // db에 쿼리문 입력						
+				pstmt2.executeUpdate(); // db에 쿼리문 입력		
+				returns = "ipModified";
 			}
 			else {	//	이미지가 없을 때
 				sql2 = "insert into ip (number, ip_image) values (?, ?)"; // ip 테이블의 이미지 넣기
-				pstmt2 = conn.prepareStatement(sql);
+				pstmt2 = conn.prepareStatement(sql2);
 				pstmt2.setInt(1, number);
 				pstmt2.setBinaryStream(2, fin, (int)imgFile.length());	//	데이터 베이스에 이미지 파일 업로드
-				pstmt2.executeUpdate(); // db에 쿼리문 입력				
+				pstmt2.executeUpdate(); // db에 쿼리문 입력	
+				returns = "ipUploaded";
 			}	
-			returns = "ipUploaded";
+			
 		} catch (Exception e) {
 			e.printStackTrace();
 			returns = "error";
@@ -183,11 +206,20 @@ public class ImageUpload {
 		try {
 			Class.forName("org.mariadb.jdbc.Driver");
 			conn = DriverManager.getConnection(dbc.getURL(), dbc.getID(), dbc.getPW()); // 데이터베이스 접근을 위한 로그인
-			sql = "delete from ip where number=?"; // ip 테이블의 number에 해당되는 레코드 삭제
+			sql = "select * from ip where number=?";
 			pstmt = conn.prepareStatement(sql);
 			pstmt.setInt(1, number);
-			pstmt.executeQuery(); // db에 쿼리문 입력
-			returns = "ipDeleted";
+			rs = pstmt.executeQuery();
+			if(rs.next()) {
+				sql2 = "delete from ip where number=?"; // organization테이블의 number에 해당되는 레코드 삭제
+				pstmt2 = conn.prepareStatement(sql2);
+				pstmt2.setInt(1, number);
+				pstmt2.executeQuery(); // db에 쿼리문 입력
+				returns = "ipDeleted";				
+			}
+			else {
+				returns = "ipNotExist";
+			}
 		} catch (Exception e) {
 			e.printStackTrace();
 			returns = "error";
@@ -212,19 +244,21 @@ public class ImageUpload {
 			rs = pstmt.executeQuery();
 			if(rs.next()) {	//	이미지가 있을 때
 				sql2 = "update rule set role_image=? where number=?"; // rule 테이블의 이미지 업데이트
-				pstmt2 = conn.prepareStatement(sql);
+				pstmt2 = conn.prepareStatement(sql2);
 				pstmt2.setBinaryStream(1, fin, (int)imgFile.length());	//	데이터 베이스에 이미지 파일 업로드
 				pstmt2.setInt(2, number);
-				pstmt2.executeUpdate(); // db에 쿼리문 입력						
+				pstmt2.executeUpdate(); // db에 쿼리문 입력	
+				returns = "ruleModified";
 			}
 			else {	//	이미지가 없을 때
 				sql2 = "insert into rule (number, role_image) values (?, ?)"; // rule 테이블의 이미지 넣기
-				pstmt2 = conn.prepareStatement(sql);
+				pstmt2 = conn.prepareStatement(sql2);
 				pstmt2.setInt(1, number);
 				pstmt2.setBinaryStream(2, fin, (int)imgFile.length());	//	데이터 베이스에 이미지 파일 업로드
-				pstmt2.executeUpdate(); // db에 쿼리문 입력				
+				pstmt2.executeUpdate(); // db에 쿼리문 입력	
+				returns = "ruleUploaded";
 			}	
-			returns = "ruleUploaded";
+
 		} catch (Exception e) {
 			e.printStackTrace();
 			returns = "error";
@@ -241,11 +275,20 @@ public class ImageUpload {
 		try {
 			Class.forName("org.mariadb.jdbc.Driver");
 			conn = DriverManager.getConnection(dbc.getURL(), dbc.getID(), dbc.getPW()); // 데이터베이스 접근을 위한 로그인
-			sql = "delete from rule where number=?"; // rule 테이블의 number에 해당되는 레코드 삭제
+			sql = "select * from rule where number=?";
 			pstmt = conn.prepareStatement(sql);
 			pstmt.setInt(1, number);
-			pstmt.executeQuery(); // db에 쿼리문 입력
-			returns = "ipDeleted";
+			rs = pstmt.executeQuery();
+			if(rs.next()) {
+				sql2 = "delete from rule where number=?"; // organization테이블의 number에 해당되는 레코드 삭제
+				pstmt2 = conn.prepareStatement(sql2);
+				pstmt2.setInt(1, number);
+				pstmt2.executeQuery(); // db에 쿼리문 입력
+				returns = "ruleDeleted";				
+			}
+			else {
+				returns = "ruleNotExist";
+			}
 		} catch (Exception e) {
 			e.printStackTrace();
 			returns = "error";
